rename sprites into player for proper naming, syntax changes
