+++ player.py
@@ -5,15 +5,15 @@ import random
 
 class Player(pygame.sprite.Sprite):
     
-    def __init__(self, game, x, y):
+    def __init__(self, game, x, y, image_list):
 
         self.game = game
         self._layer = PLAYER_LAYER
         self.groups = self.game.all_sprites
         pygame.sprite.Sprite.__init__(self, self.groups) #add player to all sprites group
 
-        #Player Lives
         self.lives = PLAYER_LIVES
+        self.last_shot_time = 0  # Initialize the last shot time
 
         # player position based on tile
         self.x = x * TILESIZE
@@ -22,15 +22,16 @@ class Player(pygame.sprite.Sprite):
         self.width = TILESIZE
         self.height = TILESIZE
 
-        # create player (TEMPORARY look)
-        self.og_image = pygame.image.load('Images/ships/ship-a/ship-a1.png').convert_alpha()
+        # create player
+        self.image_list = image_list 
+        self.og_image = self.image_list[0]
         self.image = self.og_image
-        self.damaged_image = pygame.image.load('Images/ships/ship-a/ship-a-damaged.png').convert_alpha()
+        self.damaged_image = self.image_list[3]
         self.damage_loop = 0
 
         # self.image.get_rect() returns a new rectangle covering the entire surface of `self.image`. This rectangle (rect) is used to position the sprite on the screen.
         # it's important for collision detection and rendering the sprite at its current position.
-        self.rect = self.image.get_rect()
+        self.rect = self.og_image.get_rect()
 
         # set player's rect x, y positions
         self.rect.x = self.x
@@ -45,23 +46,22 @@ class Player(pygame.sprite.Sprite):
         self.x_change = 0
         self.y_change = 0
         self.angle = 0
-
         self.player_bullets = self.game.player_bullets
-        
         
     #update player sprite
     def update(self):
-        current_time = pygame.time.get_ticks()   
 
+        current_time = pygame.time.get_ticks()           
         #update movement
         self.rotate()
         self.movement()
-        
+        #update collision check
+        #self.collide_asteroid()
+
         #check collisions
-        self.collide(self.game.ship_reg_bullets)
+        self.collide(self.game.ship_bullets)
         self.collide(self.game.asteroids)
         self.collide(self.game.ships)
-        self.collide(self.game.ship_sp_bullets)
 
         #update acceleration
         self.rect.center += self.velocity  # Apply velocity to the player's position
@@ -75,7 +75,6 @@ class Player(pygame.sprite.Sprite):
         #reset _change vars
         self.x_change = 0
         self.y_change = 0
-        
 
         # Flickering logic: Change image back and forth if within invulnerability period
         if current_time <= self.damage_loop + 3000:  # 3000 ms invulnerability
@@ -87,31 +86,40 @@ class Player(pygame.sprite.Sprite):
             self.image = self.og_image  # Outside invulnerability period, use original image
 
         self.rotate()
-
         self.handle_input()
 
     def shoot_regular_bullet(self):
-        bullet = RegularBullet(self.rect.centerx, self.rect.centery)
-        # Set velocity based on player's direction
-        bullet.vel_x = math.cos(math.radians(self.angle)) * bullet.speed
-        bullet.vel_y = -math.sin(math.radians(self.angle)) * bullet.speed
+        bullet = RegularBullet(self, self.rect.centerx, self.rect.centery, self.angle)
+        rad_angle = math.radians(self.angle)  # Convert angle to radians
+        bullet.vel_x = math.cos(rad_angle) * bullet.speed  # Calculate x velocity based on angle
+        bullet.vel_y = math.sin(rad_angle) * bullet.speed  # Calculate y velocity based on angle
         self.game.all_sprites.add(bullet)
         self.player_bullets.add(bullet)
 
     def shoot_special_bullet(self):
-        bullet = SpecialBullet(self.rect.centerx, self.rect.centery)
-        # Set velocity based on player's direction
-        bullet.vel_x = math.cos(math.radians(self.angle)) * bullet.speed
-        bullet.vel_y = -math.sin(math.radians(self.angle)) * bullet.speed
+        bullet = SpecialBullet(self.rect.centerx, self.rect.centery, self.angle)
+        rad_angle = math.radians(self.angle)  # Convert angle to radians
+        bullet.vel_x = math.cos(rad_angle) * bullet.speed  # Calculate x velocity based on angle
+        bullet.vel_y = math.sin(rad_angle) * bullet.speed  # Calculate y velocity based on angle
         self.game.all_sprites.add(bullet)
         self.player_bullets.add(bullet)
 
     def handle_input(self):
         keys = pygame.key.get_pressed()
-        if keys[pygame.K_SPACE]:
+        current_time = pygame.time.get_ticks()
+
+        # Calculate the time elapsed since the last shot
+        time_since_last_shot = current_time - self.last_shot_time
+
+        if keys[pygame.K_SPACE] and time_since_last_shot >= 500:  # Shoot only if 500 milliseconds (0.5 second) have passed since the last shot
             self.shoot_regular_bullet()  # Shoot regular bullet when space key is pressed
-        elif keys[pygame.K_LSHIFT]:
+            PLAYER_CHANNEL.play(PLAYER_BULLET_MUSIC)
+            self.last_shot_time = current_time  # Update the last shot time
+
+        elif keys[pygame.K_LSHIFT] and time_since_last_shot >= 500:
             self.shoot_special_bullet()
+            PLAYER_CHANNEL.play(PLAYER_BULLET_MUSIC)
+            self.last_shot_time = current_time  # Update the last shot time
 
     def wrap_around_screen(self):
         if self.rect.right < 0:
@@ -167,6 +175,7 @@ class Player(pygame.sprite.Sprite):
             # Check if within collision threshold and not currently invulnerable
             if distance < collision_threshold and current_time > self.damage_loop + 3000:  # Assuming 3000 ms invulnerability
                 self.lives -= 1
+                PLAYER_DESTROYED_CHANNEL.play(PLAYER_DESTROYED_MUSIC)
                 self.damage_loop = current_time  # Reset invulnerability timer
                 
                 if self.lives <= 0:
@@ -174,43 +183,79 @@ class Player(pygame.sprite.Sprite):
                     self.game.playing = False
                     break
             
-
 class RegularBullet(pygame.sprite.Sprite):
-    def __init__(self, x, y):
+    def __init__(self, game, x, y, angle):
         super().__init__()
+        self.game = game
         self.image = pygame.Surface((BULLET_SIZE, BULLET_SIZE))
         self.image.fill(BULLET_COLOR)
         self.rect = self.image.get_rect(center=(x, y))
         self.speed = BULLET_SPEED
-        self.vel_x = 0  # Velocity in the x direction
-        self.vel_y = -self.speed  # Velocity in the y direction (initially upwards)
+        self.angle = angle  # Store the angle passed from the player
+        self.vel_x = math.cos(math.radians(self.angle)) * self.speed  # Calculate x velocity based on angle
+        self.vel_y = math.sin(math.radians(self.angle)) * self.speed  # Calculate y velocity based on angle
+        self.creation_time = pygame.time.get_ticks()
 
     def update(self):
-        # Update bullet position based on velocity
         self.rect.x += self.vel_x
         self.rect.y += self.vel_y
-
-        # Remove bullet if it goes off-screen
-        if self.rect.bottom < 0 or self.rect.top > WIN_HEIGHT or self.rect.right < 0 or self.rect.left > WIN_WIDTH:
+        #self.collide(self.game.asteroids)
+        # destroy bullet exists for more than 2 seconds
+        if pygame.time.get_ticks() - self.creation_time > 2000: 
             self.kill()
 
+        # leaves the screen = reenters from the opposite side
+        if self.rect.bottom < 0: 
+            self.rect.y = WIN_HEIGHT
+            self.rect.x = WIN_WIDTH - self.rect.x
+        if self.rect.right < 0:
+            self.rect.y = WIN_HEIGHT - self.rect.y
+            self.rect.x = WIN_WIDTH
+        if self.rect.top > WIN_HEIGHT:
+            self.rect.y = 0
+            self.rect.x = WIN_WIDTH - self.rect.x
+        if self.rect.left > WIN_WIDTH:
+            self.rect.y = WIN_HEIGHT - self.rect.y
+            self.rect.x = 0
+
 class SpecialBullet(pygame.sprite.Sprite):
-    def __init__(self, x, y):
+    def __init__(self, x, y, angle):
         super().__init__()
         self.image = pygame.Surface((BULLET_SIZE, BULLET_SIZE))
         self.image.fill(SPECIAL_BULLET_COLOR)
         self.rect = self.image.get_rect(center=(x, y))
         self.speed = SPECIAL_BULLET_SPEED
-        self.vel_x = 0  # Velocity in the x direction
-        self.vel_y = -self.speed  # Velocity in the y direction (initially upwards)
+        self.angle = angle  # store the angle passed from the player
+        self.vel_x = math.cos(math.radians(self.angle)) * self.speed  # calculate x velocity based on angle
+        self.vel_y = math.sin(math.radians(self.angle)) * self.speed  # calculate y velocity based on angle
+        self.creation_time = pygame.time.get_ticks()
 
     def update(self):
-        # Update bullet position based on velocity
         self.rect.x += self.vel_x
         self.rect.y += self.vel_y
-
-        # Remove bullet if it goes off-screen
-        if self.rect.bottom < 0 or self.rect.top > WIN_HEIGHT or self.rect.right < 0 or self.rect.left > WIN_WIDTH:
+        
+        # destroy bullet exists for more than 2 seconds
+        if pygame.time.get_ticks() - self.creation_time > 2000: 
             self.kill()
+
+        # leaves the screen = reenters from the opposite side
+        if self.rect.bottom < 0: 
+            self.rect.y = WIN_HEIGHT
+            self.rect.x = WIN_WIDTH - self.rect.x
+        if self.rect.right < 0:
+            self.rect.y = WIN_HEIGHT - self.rect.y
+            self.rect.x = WIN_WIDTH
+        if self.rect.top > WIN_HEIGHT:
+            self.rect.y = 0
+            self.rect.x = WIN_WIDTH - self.rect.x
+        if self.rect.left > WIN_WIDTH:
+            self.rect.y = WIN_HEIGHT - self.rect.y
+            self.rect.x = 0
+
+    def collide(self, spriteGroup):
+        for asteroid in spriteGroup:
+            if pygame.sprite.collide_circle(self, asteroid):
+                self.kill()
+                asteroid.take_damage()
